requirements.txt
asgiref==3.3.1
beautifulsoup4
Django==3.1.5
django-countries==7.0
#plotly==4.14.3
#pandas==1.2.2
psycopg2==2.8.6
requests==2.25.1
gevent==20.12.1
gunicorn[gevent]==20.0.4
WeasyPrint==52.5
django-markdownify==0.9.0
